feat: add redirect link
